requirements.txt
streamlit==1.56.0
faiss-cpu==1.13.2
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.0.7
google-generativeai==0.8.6
pypdf==6.9.2
python-dotenv==1.1.1
uvicorn
serpapi
google-search-results
youtube-search
youtube-transcript-api
bm25
